requirements.txt
requests
pandas
python-dotenv
openai
tableauhyperapi
tableauserverclient
